Fixed Issue 26.
Added a new listener (ISelectionListener as a postSelectionListener) to handle both cursor change events and also editor file change events.
Added one new observer pattern: CursorChangedListener and CursorChangedNotifier.
SynchronizerPartListener (which was implementing IPartListener2 to get editor file change events), ActiveFileChangedNotifier and ActiveFileChangedListener are now deprecated. For now, I still have the code, but in the future they will probably get deleted.

Current order of computing compilation errors:
1- Compilation errors on the current editor file with respect to their distance from the current cursor location (distance is in terms of offset, not pixel distance)
2- Other compilation error files.

I realized that we are still missing the difference between compilation errors on files that are open in Eclipse editor in the background vs. files that are completely closed.